Remove unused adapters, tidy up others
