+++ packages/protvista-feature-adapter/README.md
@@ -1,4 +1,1 @@
 # protvista-feature-adapter
-A specialization of the uniprot-entry- data-adapter for ProtVista features (regions and sites)
-
-See it running [Here](https://ebi-webcomponents.github.io/protvista-feature-adapter/). Remember to open the console to see the loading.
