--- packages/protvista-feature-adapter/examples/index.html
@@ -1,50 +1,0 @@
-<!doctype html>
-<html class="no-js" lang="">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>protvista-feature-adapter-demo</title>
-
-</head>
-<body>
-<protvista-feature-adapter id="adapter1">
-    <data-loader id="loader1">
-        <source src="https://www.ebi.ac.uk/proteins/api/features/P05067" />
-    </data-loader>
-</protvista-feature-adapter>
-
-<!-- Required for IE11 -->
-<script src="../../../node_modules/@babel/polyfill/dist/polyfill.min.js"></script>
-<!-- Web component polyfill (only loads what it needs) -->
-<script src="../../../node_modules/@webcomponents/webcomponentsjs/webcomponents-lite.js" charset="utf-8"></script>
-<!-- Required to polyfill modern browsers as code is ES5 for IE... -->
-<script src="../../../node_modules/@webcomponents/webcomponentsjs/custom-elements-es5-adapter.js" charset="utf-8"></script>
-
-<script src="../../protvista-uniprot-entry-adapter/dist/protvista-uniprot-entry-adapter.js" charset="utf-8" defer></script>
-<script src="../../data-loader/dist/data-loader.js" charset="utf-8" defer></script>
-
-<script src="../dist/protvista-feature-adapter.js" charset="utf-8" defer></script>
-
-<script>
-    for (const adapter of document.querySelectorAll('protvista-feature-adapter')) {
-      adapter.addEventListener(
-        'load',
-        e => {
-          console.log('Data loaded');
-          console.log(e.target);
-          console.log(e.detail);
-        }
-      );
-      adapter.addEventListener(
-        'error',
-        e => {
-          console.log('error');
-          console.log(e.target);
-          console.log(e.detail);
-        }
-      );
-    }
-</script>
-
-</body>
-</html>
