+++ packages/protvista-feature-adapter/src/index.js
@@ -1,5 +1,10 @@
-import ProtvistaFeatureAdapter from "./ProtVistaFeatureAdapter";
+import ProtvistaFeatureAdapter from "./ProtvistaFeatureAdapter";
 
 if (window.customElements) {
-  customElements.define("protvista-feature-adapter", ProtvistaFeatureAdapter);
+  customElements.define(
+    "protvista-uniprot-entry-adapter",
+    ProtvistaFeatureAdapter
+  );
 }
+
+export default ProtvistaFeatureAdapter;
