--- packages/protvista-topology-adapter/README.md
@@ -1,2 +1,0 @@
-# protvista-topology-adapter
-A dataloader and parser (ProtVista compliant) for topological features in UniProt
