--- packages/protvista-topology-adapter/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "protvista-topology-adapter",
-  "description": "A dataloader and parser (ProtVista compliant) for topological features in UniProt",
-  "main": "dist/protvista-topology-adapter.js",
-  "module": "src/ProtVistaTopologyAdapter.js",
-  "files": [
-    "src",
-    "dist"
-  ],
-  "keywords": [
-    "ProtVista",
-    "dataloader",
-    "parser",
-    "topology"
-  ],
-  "version": "2.1.12",
-  "gitHead": "5a756316482fa31d085602d46438861fb45e5178",
-  "dependencies": {
-    "lodash-es": "^4.17.11"
-  }
-}
