--- packages/protvista-topology-adapter/src/index.js
@@ -1,5 +1,0 @@
-import ProtVistaTopologyAdapter from "./ProtVistaTopologyAdapter";
-
-if (window.customElements) {
-  customElements.define("protvista-topology-adapter", ProtVistaTopologyAdapter);
-}
