--- packages/protvista-uniprot-entry-adapter/README.md
@@ -1,6 +1,0 @@
-# protvista-uniprot-entry-adapter
-A generic/abstract adapter for services retrieving UniProt entries by accession
-
-Not intended to be used on its own but to be extended by specialized adapters such as the protvista -feature-adapter
-
-See it running [Here](https://ebi-webcomponents.github.io/protvista-uniprot-entry-adapter/). Remember to open the console to see the loading.
