--- packages/protvista-uniprot-entry-adapter/index.html
@@ -1,52 +1,0 @@
-<!doctype html>
-<html class="no-js" lang="">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>protvista-uniprot-entry-adapter-demo</title>
-
-</head>
-<body>
-    <protvista-uniprot-entry-adapter>
-        <data-loader id="loader1" accession="P05067">
-            <source src="https://www.ebi.ac.uk/proteins/api/proteins/P05067" />
-        </data-loader>
-        <data-loader  id="additionalLoader">
-            <source src="https://www.ebi.ac.uk/proteins/api/proteins/P05067" />
-        </data-loader>
-    </protvista-uniprot-entry-adapter>
-
-    <!-- Required for IE11 -->
-    <script src="https://cdn.jsdelivr.net/npm/babel-polyfill/dist/polyfill.min.js"></script>
-    <!-- Web component polyfill (only loads what it needs) -->
-    <script src="https://cdn.jsdelivr.net/npm/@webcomponents/webcomponentsjs/webcomponents-lite.js" charset="utf-8"></script>
-    <!-- Required to polyfill modern browsers as code is ES5 for IE... -->
-    <script src="https://cdn.jsdelivr.net/npm/@webcomponents/webcomponentsjs/custom-elements-es5-adapter.js" charset="utf-8"></script>
-
-    <script src="https://cdn.jsdelivr.net/npm/protvista-uniprot-entry-adapter/dist/ProtVistaUniProtEntryAdapter.js" charset="utf-8" defer></script>
-    <script src="https://cdn.jsdelivr.net/npm/data-loader@latest/dist/index.js" charset="utf-8" defer></script>
-
-
-    <script>
-        for (const adapter of document.querySelectorAll('protvista-uniprot-entry-adapter')) {
-          adapter.addEventListener(
-            'load',
-            e => {
-              console.log('Data loaded');
-              console.log(e.target);
-              console.log(e.detail);
-            }
-          );
-          adapter.addEventListener(
-            'error',
-            e => {
-              console.log('error');
-              console.log(e.target);
-              console.log(e.detail);
-            }
-          );
-        }
-    </script>
-
-</body>
-</html>
