--- packages/protvista-uniprot-entry-adapter/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "protvista-uniprot-entry-adapter",
-  "version": "2.1.12",
-  "description": "A generic/abstract adapter for services retrieving UniProt entries by accession",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "main": "dist/protvista-uniprot-entry-adapter.js",
-  "module": "src/ProtVistaUniProtEntryAdapter.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ebi-uniprot/protvista-uniprot-entry-adapter.git"
-  },
-  "keywords": [
-    "ProtVista",
-    "dataloader",
-    "parser",
-    "protein-structure"
-  ],
-  "author": "ljgarcia",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/ebi-uniprot/protvista-uniprot-entry-adapter/issues"
-  },
-  "gitHead": "5a756316482fa31d085602d46438861fb45e5178",
-  "dependencies": {
-    "lodash-es": "^4.17.11"
-  }
-}
