--- packages/protvista-uniprot-entry-adapter/src/evidences.js
@@ -1,63 +1,0 @@
-export const ecoMap = [{
-    "name": "ECO:0000269",
-    "description": "Manual assertion based on experiment",
-    "shortDescription" : "Publication",
-    "acronym": "EXP",
-    "isManual": true
-}, {
-    "name": "ECO:0000303",
-    "description": 'Manual assertion based on opinion',
-    "shortDescription" : "Publication",
-    "acronym": "NAS",
-    "isManual": true
-}, {
-    "name": "ECO:0000305",
-    "description": 'Manual assertion inferred by curator',
-    "shortDescription" : "Curated",
-    "acronym": "IC",
-    "isManual": true
-}, {
-    "name": "ECO:0000250",
-    "description": "Manual assertion inferred from sequence similarity",
-    "shortDescription" : "By similarity",
-    "acronym": "ISS",
-    "isManual": true
-}, {
-    "name": "ECO:0000255",
-    "description": "Manual assertion according to rules",
-    "shortDescription" : "Sequence Analysis",
-    "acronym": "ISM",
-    "isManual": true
-}, {
-    "name": "ECO:0000244",
-    "description": "Manual assertion inferred from combination of experimental and computational evidence",
-    "shortDescription" : "Combined sources",
-    "acronym": "MIXM",
-    "isManual": true
-}, {
-    "name": "ECO:0000312",
-    "description": "Manual assertion inferred from database entries",
-    "shortDescription" : "Imported",
-    "acronym": "MI",
-    "isManual": true
-}, {
-    "name": "ECO:0000256",
-    "description": "Automatic assertion according to rules",
-    "shortDescription" : "Automatic annotation",
-    "acronym": "AA"
-}, {
-    "name": "ECO:0000259",
-    "description": "Automatic assertion inferred from signature match",
-    "shortDescription" : "Automatic annotation",
-    "acronym": "AA"
-}, {
-    "name": "ECO:0000213",
-    "description": "Automatic assertion inferred from combination of experimental and computational evidence",
-    "shortDescription" : "Combined sources",
-    "acronym": "MIXA"
-}, {
-    "name": "ECO:0000313",
-    "description": "Automatic assertion inferred from database entries",
-    "shortDescription" : "Imported",
-    "acronym": "AI"
-}];
