--- packages/protvista-uniprot-entry-adapter/src/index.js
@@ -1,10 +1,0 @@
-import ProtvistaUniprotEntryAdapter from "./ProtVistaUniProtEntryAdapter";
-
-if (window.customElements) {
-  customElements.define(
-    "protvista-uniprot-entry-adapter",
-    ProtvistaUniprotEntryAdapter
-  );
-}
-
-export default ProtvistaUniprotEntryAdapter;
